Lab2/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
python-multipart
opencv-python
pillow
python-jose
